Simple :trollface:

+++ lib/tasks/translations/transifex_updater.rb
@@ -31,19 +31,69 @@ class TransifexUpdater
 
   # Processes each new translation file (if any) and applies updates to original
   def update_translations
-    new_translations = Dir.glob(Rails.root.join("config/locales/*.new"))
+    return unless new_translation_files.any?
 
-    new_translations.each do |new_file|
-      existing_file = new_file.chomp('.new')
+    new_translation_files.each do |new_file|
+      current_file = new_file.chomp('.new')
+      updated_translations = apply_updates(current_file, new_file)
 
-      merged_data = deep_diff_merge(YAML.load_file(existing_file), YAML.load_file(new_file))
-
-      File.open(existing_file, "w") do |file|
-        file.write merged_data.to_yaml(line_width: -1)
+      File.open(current_file, "w") do |file|
+        file.write updated_translations.to_yaml(line_width: -1)
       end
 
       File.delete(new_file)
     end
+  end
+
+  def apply_updates(current_file, new_file)
+    current_translations = YAML.load_file(current_file)
+    upstream_translations = YAML.load_file(new_file)
+
+    puts "\n####### Locale: #{root_key(current_translations)} #######\n\n"
+
+    puts "####### New compatible keys in basefile:\n\n"
+    updated_structure = { root_key(current_translations) =>
+                              merge_keys(keys(current_translations), keys(base_translations)) }
+
+    puts "\n####### New/updated compatible translations:\n\n"
+    merge_values(updated_structure, upstream_translations)
+  end
+
+  # The root key in any locale, eg: "fr"
+  def root_key(translations)
+    translations.first[0]
+  end
+
+  # Everything else underneath the root_key
+  def keys(translations)
+    translations.first[1]
+  end
+
+  def new_translation_files
+    Dir.glob(Rails.root.join("config/locales/*.new"))
+  end
+
+  def base_translations
+    YAML.load_file(Rails.root.join("config/locales/en.yml"))
+  end
+
+  # Recursively diffs and reverse-merges keys from second_hash into first_hash, where both
+  # hashes are deeply-nested datasets of unknown depth (eg converted from complex YAML).
+  # - Keys missing in second_hash are not removed from first_hash
+  # - Keys present in second_hash that are not present in first_hash are added to first_hash
+  def merge_keys(first_hash, second_hash)
+    second_hash.each_pair do |second_key, second_value|
+      first_value = first_hash[second_key]
+
+      if second_value.is_a?(Hash) && first_value.is_a?(Hash)
+        first_hash[second_key] = merge_keys(first_value, second_value)
+      elsif !first_hash.key? second_key
+        puts JSON.pretty_generate({second_key => second_value})
+        first_hash[second_key] = second_value
+      end
+    end
+
+    first_hash
   end
 
   # Recursively diffs and reverse-merges values from second_hash into first_hash, where both
@@ -52,15 +102,16 @@ class TransifexUpdater
   # - Keys present in second_hash that are not present in first_hash are not added
   # - Values in second_hash overwrite values in first_hash, if the keys are present
   #   in both hashes, and a value is present (non-blank) in second_hash
-  def deep_diff_merge(first_hash, second_hash)
+  def merge_values(first_hash, second_hash)
     first_hash.each_pair do |current_key, current_value|
       new_value = second_hash[current_key]
 
-      first_hash[current_key] = if current_value.is_a?(Hash) && new_value.is_a?(Hash)
-                                  deep_diff_merge current_value, new_value
-                                else
-                                  new_value.present? ? new_value : current_value
-                                end
+      if current_value.is_a?(Hash) && new_value.is_a?(Hash)
+        first_hash[current_key] = merge_values(current_value, new_value)
+      elsif new_value.present? && new_value != current_value
+        puts "#{current_key}: #{new_value}"
+        first_hash[current_key] = new_value
+      end
     end
 
     first_hash
